Add basic methods to GAPNode

+++ EP2300-Makeup-CodeTemplate/src/peersim/EP2300/util/NodeStateVector.java
@@ -3,9 +3,9 @@ package peersim.EP2300.util;
 public class NodeStateVector {
 	public String status;
 	public double level;
-	public float aggregate;
+	public long aggregate;
 
-	public NodeStateVector(String status, double level, float aggregate) {
+	public NodeStateVector(String status, double level, long aggregate) {
 		this.status = status;
 		this.level = level;
 		this.aggregate = aggregate;
